Unify ParallelGrids, add parallel grid groups, remove amplitude mode.
================================

Main computations call sequence:

performSteps          <- main entry point to start of computations (both sequential, parallel and gpu)
  |                      (+ saves result)
  |
  |
performNSteps         <- performs computations for specific number of steps and does required
  =                      share operations
  =
  =
performNStepsXD       <- same as performNStep but for specified dim: does computations for all blocks
  |                      one after another and performs share operations (afterwards for gpu and
  |                      in-process for parallel mode) + does rebalance and saves intermediate results
  |
performNStepsForBlock <- performs computations for specific block for specified number of steps
  |                      (this includes copying data to/from gpu, share operations for simple parallel
  |                      mode)
 ...

shareE function is used in two separate places, on for in-process share operations for simple parallel mode (in performNStepsForBlock), another one for gpu+parallel share ops (in performNStepsXD). This is done for efficiency (not to copy data for sharing on half time step), but this requires buffer size to be at least 2, because number of time steps will be -1.

=============================

Parallel entities dependencies:

    Scheme
      |______________________
                            |
                            |
    _ParallelGroup1         |        ParallelGroup2
   /   ^                    |          ^           \
  /    |_____________       |          |            \
  |    |             \      |          |            |
  |    |              \     v          |            |
  |  ParallelGrid1   ParallelGrid2   ParallelGrid3  |
  |    |              /               /             |
  |    |_____________/_______________/              |
  \    v                                           /
   \>ParallelGridCore<____________________________/